Moving the stylelint config file to config/
