+++ lib/config/stylelint.js
