+++ lib/lint_css.js
@@ -1,6 +1,6 @@
 var _ = require('lodash');
 var stylelint = require('stylelint');
-var STYLELINT_CONFIG = require('./stylelint_config');
+var STYLELINT_CONFIG = require('./config/stylelint');
 var glob = require('glob');
 var path = require('path');
